openssl: Add all patches as sources for tito
Tito does not copy patches into its temporary %_specdir, which causes
the build with tito to fail. Since we're just using tito as a glorified
way to download an upstream tarball before the build, this isn't the
behavior we want.

Fortunately, we can work around it by mentioning all patch files as
sources, and a bit of regex magic makes this simple to do.

See https://github.com/rpm-software-management/tito/issues/446.

Signed-off-by: Clemens Lang <cllang@redhat.com>